fastapi-gateway/pyproject.toml
[project]
name = "fastapi-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "pyjwt[crypto]>=2.10.1",
    "protobuf>=6.31.1",
    "uvicorn>=0.41.0",
    "fastapi-limiter>=0.2.0",
]
